refactor interactive_process intrinsic
